requirements.txt
requests
beautifulsoup4
pandas
sentence-transformers
chromadb
google-generativeai
fastapi
uvicorn[standard]
python-dotenv
streamlit